Allow `list[T]` to extend tokens in ListPlaceNode of type `T`
Previously only exact matches of the returned value from a `FunctionTransitionNode` function call and the type of the
destination `ListPlaceNode` node were allowed. The change here is that if a function returns a list of items of type
`T`, and the destination `ListPlaceNode` is of type `T`, then the items in the list are added to the tokens of the
`ListPlaceNode`.

 - Create toy example of this in `01_fill_place.ipynb`.
 - Add commonly used visualisation code to `plotting/simple_graphviz.py`.
 - Allow `ExecutableGraph` nodes to be instantiated using positional arguments as well as keyword arguments.